requirements.txt
Django>=5.2.5
djangorestframework>=3.14
django-filter>=23.1
psycopg2-binary>=2.9
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.2.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
django-rest-passwordreset>=1.3.0